Look at effect of projection on delta-variance
Taking a thin sheet (LOS depth of 16 pixels out of 256) from a fBM
cube does seem to reduce the slope of the delta-variance, but we don't
see a clear break in the spectrum at the scale of the LOS depth.  I
need to investigate more the variation with power spectrum slope,
k. Also, I should do the structure function for the thin sheets
too. And calculate the spatial power spectrum for good measure.

There are also some interesting trends in the distribution of LOS
sigma and its relation to the POS sigma. For shallow power spectra (k
< 3) we see that the LOS sigma is a rather narrow distribution around
1 for the thick cloud, and is somewhat broader and around 0.8 for the
thin sheet. At the same time, the distribution of centroid velocities
becomes significantly narrower than in the full cube, especially for
the thick cloud, but to some extent also for the thin sheet.